+++ src/ContainerRegistry/ContainerRegistry/ChangeLog.md
@@ -18,6 +18,11 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated Azure.Core to 1.34.0.
+
+## Version 4.1.0
+* Updated Azure.Core to 1.33.0.
+* Added new cmdlet `New-AzContainerRegistryCredentials`
 
 ## Version 4.0.0
 * Updated module to autorest based